kusddrphy: release {I,O}{DELAY,SERDES}E3.RST with IDELAYCTRL.RST
https://forums.xilinx.com/t5/UltraScale-Architecture/Reset-sequencing-for-IDELAYCTRL-and-IDELAYE3-is-contradictory-in/td-p/773938

There is a difference to previous families that if you hold the IDELAY
reset high while the IDELAYCTRL is high then the IDELAYCTRL will not be
able to calibrate properly. In 7 Series it was a normal follow for the
user to release the reset of the IDELAYCTRL wait for RDY and then
release the IDELAY reset. That is changed for UltraScale. They are both
correct, what the table on p167 says is

The RST pin (reset) is synchronous with the CLK. When the IDELAYE3 is
reset, the delay is set to the value defined by the DELAY_VALUE
attribute. The reset of the IDELAYE3 should not be asserted until after
the initial reset release sequence and
IDELAYCTRL.RDY goes High.

So the initial reset sequence says :

d. Release the reset of following primitives: IDELAYCTRL, IDELAY,
ISERDES, OSERDES,
e. Wait until the RDY of all the used IDELAYCTRL primitives are
asserted High.
Now the application in the FPGA logic can be released after a delay of
at least 64 clock cycles.

After this is done the user can reset the IDELAYE3 again while the
design is running to value in the DELAY_VALUE.